moved AntiEntropyStoreConfig into separate file, also made apply taking config param and provided a function that sets the config to default
